+++ KalikoCMS.Admin/Admin/Content/PropertyType/ImagePropertyEditor.js
@@ -1,4 +1,6 @@
 ﻿(function (propertyEditor) {
+    var localTop = top;
+
     propertyEditor.image = propertyEditor.image || {
         openDialog: function (pathField, previewImage, originalPathField, cropXField, cropYField, cropWField, cropHField, width, height, descriptionField) {
             var imagePath = $(pathField).val();
@@ -25,18 +27,21 @@
                     $(originalPathField).val('');
                 }
             };
-            
-            top.registerCallback(callback);
 
-            top.propertyEditor.dialogs.openEditImageDialog(imagePath, originalPath, cropX, cropY, cropW, cropH, width, height, description);
+
+            localTop.registerCallback(callback);
+
+            localTop.propertyEditor.dialogs.openEditImageDialog(imagePath, originalPath, cropX, cropY, cropW, cropH, width, height, description);
         }
     };
 })(top.propertyEditor || (top.propertyEditor = {}));
 
 
 (function (dialogs) {
+    var localParent = parent;
+
     dialogs.openEditImageDialog = dialogs.openEditImageDialog || function (imagePath, originalPath, cropX, cropY, cropW, cropH, width, height, description) {
-        parent.openModal("Content/Dialogs/EditImageDialog.aspx?imagePath=" + imagePath + "&originalPath=" + originalPath + "&cropX=" + cropX + "&cropY=" + cropY + "&cropW=" + cropW + "&cropH=" + cropH + "&width=" + width + "&height=" + height + "&description=" + escape(description), 710, 500);
+        localParent.openModal("Content/Dialogs/EditImageDialog.aspx?imagePath=" + imagePath + "&originalPath=" + originalPath + "&cropX=" + cropX + "&cropY=" + cropY + "&cropW=" + cropW + "&cropH=" + cropH + "&width=" + width + "&height=" + height + "&description=" + escape(description), 710, 500);
         return false;
     };
 })(top.propertyEditor.dialogs || (top.propertyEditor.dialogs = {}));