+++ ReleaseNotes.md
@@ -3,6 +3,7 @@
 * Fixed problem with omitted content type when returning MVC action result
 * Fixed problem with paths starting with "./" for images in HTML editor
 * Fixed problem with sort settings not being applied for site root in admin
+* Fixed problem with reopening image and link dialogs in admin
 * Updated jsTree to a version that not relies on registerElement which will be deprecated in Chrome
 * Validate password before trying to change when editing a user in admin
 * Added optional configuration setting (enableSessions) to support the use of session state in MVC